ci(trivy): disable vulnerability scan on PRs
There is a risk with this approach to miss a new vulnerability being introduced by a PR. But in our case, it is much less likely than a random CVE popping up in existing dependencies at the wrong time, blocking a totally unrelated PR merge. And with the daily scheduled scan, we will catch all vulnerabilities spotted by Trivy soon enough (certainly way before I include the PR in a release).